iread.table {iterators}	R Documentation

Iterator over Rows of a Data Frame Stored in a File

Description

Returns an iterator over the rows of a data frame stored in a file in table format. It is a wrapper around the standard read.table function.

Usage

iread.table(file, ..., verbose=FALSE)

Arguments

file
the name of the file to read the data from.

...
all additional arguments are passed on to the read.table function. See the documentation for read.table for more information.

verbose
logical value indicating whether or not to print the calls to read.table.

Value

The file reading iterator.

Note

In this version of iread.table, both the read.table arguments header and row.names must be specified. This is because the default values of these arguments depend on the contents of the beginning of the file. In order to make the subsequent calls to read.table work consistently, the user must specify those arguments explicitly. A future version of iread.table may remove this requirement.

See Also

read.table

[Package iterators version 1.0.14 Index]