ireadLines {iterators}	R Documentation

Iterator over Lines of Text from a Connection

Description

Returns an iterator over the lines of text from a connection. It is a wrapper around the standard readLines function.

Usage

ireadLines(con, n=1, ...)

Arguments

con
a connection object or a character string.

n
integer. The maximum number of lines to read. Negative values indicate that one should read up to the end of the connection. The default value is 1.

...
passed on to the readLines function.

Value

The line reading iterator.

See Also

readLines

Examples

  # create an iterator over the lines of COPYING
  it <- ireadLines(file.path(R.home(), 'COPYING'))
  nextElem(it)
  nextElem(it)
  nextElem(it)

[Package iterators version 1.0.14 Index]