iterators-package {iterators}	R Documentation

The Iterators Package

Description

The iterators package provides tools for iterating over various R data structures. Iterators are available for vectors, lists, matrices, data frames, and files. By following very simple conventions, new iterators can be written to support any type of data source, such as database queries or dynamically generated data.

Details

Further information is available in the following help topics:

iter 	 Generic function used to create iterator objects.

nextElem 	 Generic function used to get the next element of a iterator.

icount 	 A function used to create a counting iterator.

idiv 	 A function used to create a number dividing iterator.

ireadLines 	 A function used to create a file reading iterator.

For a complete list of functions with individual help pages, use library(help="iterators").

[Package iterators version 1.0.14 Index]