isplit {iterators}	R Documentation

Split Iterator

Description

Returns an iterator that divides the data in the vector x into the groups defined by f.

Usage

isplit(x, f, drop=FALSE, ...)

Arguments

x
vector or data frame of values to be split into groups.

f
a factor or list of factors used to categorize x.

drop
logical indicating if levels that do not occur should be dropped.

...
current ignored.

Value

The split iterator.

See Also

split

Examples

x <- rnorm(200)
f <- factor(sample(1:10, length(x), replace=TRUE))

it <- isplit(x, f)
expected <- split(x, f)

for (i in expected) {
  actual <- nextElem(it)
  stopifnot(actual$value == i)
}

[Package iterators version 1.0.7 Index]